Graph_lib

+++ Graph_lib/Graph.cpp
@@ -15,7 +15,7 @@ namespace Graph_lib {
 void Shape::draw_lines() const
 {
   if (color().visibility() && 1 < points.size()) // draw sole pixel?
-    for (unsigned i = 1; i < points.size(); ++i)
+    for (size_t i = 1; i < points.size(); ++i)
       fl_line(points[i - 1].x, points[i - 1].y, points[i].x, points[i].y);
 }
